parent_domain: Set cookie on "."-prefix only
See: https://github.com/Pylons/pyramid/pull/1028#issuecomment-21523972

This makes parent_domain only set a cookie on the "."-prefixed parent domain (or
the "."-prefix of the current domain, if we're on the parent domain).  This
prevents issues where a user logs in on the parent domain, accesses a subdomain,
then logs out on the parent domain.  A cookie still exists on the subdomain, so
the user is still logged in on the subdomain when they shouldn't be.